+++ modules/core/misc/objc/src/DoubleVector.mm
@@ -65,7 +65,7 @@
 }
 
 -(double)get:(NSInteger)index {
-    if (index < 0 || index >= v.size()) {
+    if (index < 0 || index >= (long)v.size()) {
         @throw [NSException exceptionWithName:NSRangeException reason:@"Invalid data length" userInfo:nil];
     }
     return v[index];
